[Refactor] Move pipeline metrics to ExecPrimitive
