+++ be/src/exec/pipeline/pipeline_driver_executor.cpp
@@ -22,7 +22,7 @@
 #include "common/system/master_info.h"
 #include "common/thread/thread.h"
 #include "exec/pipeline/fragment_context.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/schedule/event_scheduler.h"
 #include "exec/workgroup/work_group.h"
