+++ be/src/exec/pipeline/pipeline_driver_queue.cpp
@@ -15,7 +15,7 @@
 #include "exec/pipeline/pipeline_driver_queue.h"
 
 #include "common/config_exec_flow_fwd.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/schedule/utils.h"
 #include "exec/pipeline/source_operator.h"
 #include "exec/workgroup/work_group.h"
