+++ be/src/exec/pipeline/primitives/pipeline_metrics.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 
 #include <numeric>
 #include <utility>
