+++ be/src/exec/spill/executor.h
@@ -26,6 +26,7 @@
 #include "exec/workgroup/scan_executor.h"
 #include "exec/workgroup/scan_task.h"
 #include "exec/workgroup/scan_task_queue.h"
+#include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_fwd.h"
 #include "gen_cpp/Types_types.h"
 #include "runtime/current_thread.h"
