+++ be/src/runtime/exec_env.cpp
@@ -55,7 +55,7 @@
 #include "connector/connector_registry.h"
 #include "connector/connector_sink_executor.h"
 #include "exec/pipeline/pipeline_driver_executor.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/pipeline/query_context.h"
 #include "exec/pipeline/query_context_manager.h"
 #include "exec/query_cache/cache_manager.h"
