+++ be/test/agent/resource_group_usage_recorder_test.cpp
@@ -17,7 +17,7 @@
 #include "base/testutil/assert.h"
 #include "common/config_runtime_fwd.h"
 #include "common/system/cpu_info.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_manager.h"
 #include "gtest/gtest.h"
