+++ be/test/exec/workgroup/pipeline_executor_set_manager_test.cpp
@@ -20,7 +20,7 @@
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
 #include "exec/pipeline/pipeline_driver_executor.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/workgroup/pipeline_executor_set.h"
 #include "exec/workgroup/work_group.h"
 #include "exec/workgroup/work_group_manager.h"
