+++ be/test/exec/workgroup/scan_task_queue_test.cpp
@@ -24,7 +24,7 @@
 #include "base/testutil/assert.h"
 #include "base/testutil/parallel_test.h"
 #include "exec/pipeline/pipeline_fwd.h"
-#include "exec/pipeline/pipeline_metrics.h"
+#include "exec/pipeline/primitives/pipeline_metrics.h"
 #include "exec/workgroup/priority_scan_task_queue.h"
 #include "exec/workgroup/scan_executor.h"
 #include "exec/workgroup/work_group.h"
